+++ R/form_detail.R
@@ -6,6 +6,7 @@
 #' @template param-fid
 #' @template param-url
 #' @template param-auth
+#' @template param-retries
 #' @return A tibble with one row and all form metadata as columns.
 # nolint start
 #' @seealso \url{https://odkcentral.docs.apiary.io/#reference/forms-and-submissions/'-individual-form}
@@ -44,7 +45,8 @@ form_detail <- function(pid = get_default_pid(),
                         fid = get_default_fid(),
                         url = get_default_url(),
                         un = get_default_un(),
-                        pw = get_default_pw()) {
+                        pw = get_default_pw(),
+                        retries = get_retries()) {
   yell_if_missing(url, un, pw, pid = pid, fid = fid)
   httr::RETRY(
     "GET",
@@ -58,7 +60,8 @@ form_detail <- function(pid = get_default_pid(),
       "Accept" = "application/xml",
       "X-Extended-Metadata" = "true"
     ),
-    httr::authenticate(un, pw)
+    httr::authenticate(un, pw),
+    times = retries
   ) %>%
     yell_if_error(., url, un, pw) %>%
     httr::content(.) %>%