+++ R/form_list.R
@@ -5,6 +5,7 @@
 #' @template param-pid
 #' @template param-url
 #' @template param-auth
+#' @template param-retries
 #' @return A tibble with one row per form and all form metadata as columns.
 # nolint start
 #' @seealso \url{https://odkcentral.docs.apiary.io/#reference/forms-and-submissions/forms}
@@ -36,7 +37,8 @@
 form_list <- function(pid = get_default_pid(),
                       url = get_default_url(),
                       un = get_default_un(),
-                      pw = get_default_pw()) {
+                      pw = get_default_pw(),
+                      retries = get_retries()) {
   yell_if_missing(url, un, pw, pid = pid)
   httr::RETRY(
     "GET",
@@ -45,7 +47,8 @@ form_list <- function(pid = get_default_pid(),
       "Accept" = "application/xml",
       "X-Extended-Metadata" = "true"
     ),
-    httr::authenticate(un, pw)
+    httr::authenticate(un, pw),
+    times = retries
   ) %>%
     yell_if_error(., url, un, pw) %>%
     httr::content(.) %>%