+++ man-roxygen/param-retries.R
@@ -1,0 +1,3 @@
+#' @param retries The number of attempts to retrieve a web resource.
+#'   This parameter is given to \code{\link[httr]{RETRY}(times = retries)}.
+#'   Default: 3.